Refactorisation et modularisation partielle (Travail en cours)
- Simplification du point d'entrée : suppression de l'importation récursive de main, maintenant définie localement dans main.py
- Ajout du point d'entrée dans pyproject.toml (piledger.main:main)
- Fichier __init__.py vidé (présence du module uniquement)
- Réécriture de read_data_file avec csv.reader et un gestionnaire de contexte (with)
- Remplacement de la boucle while par une boucle for dans calculate_balance pour plus de clarté
- Création du module data_handler.py pour l'import/export des données
- Séparation des fonctions liées aux transactions dans transaction.py
- Création de account.py pour les fonctions liées aux comptes
