--- src/piledger/__init__.py
@@ -1,4 +1,0 @@
-import piledger.main as piledger
-
-def main() -> None:
-    piledger.main()
